Added Car service Functionality

+++ src/main/java/DAO/CarDAO.java
@@ -51,7 +51,7 @@ public class CarDAO {
                 String color = rs.getString("color");
                 boolean cleanTitle = rs.getBoolean("clean_title");
                 int licenseNum = rs.getInt("license_num");
-                double price = rs.getInt("price");
+                double price = rs.getDouble("price");
                 Car dbCar = new Car(vin, dbMake, dbModel, modelYear, color, cleanTitle, licenseNum, price);
                 carList.add(dbCar);
             }
@@ -75,7 +75,7 @@ public class CarDAO {
                 String color = rs.getString("color");
                 boolean cleanTitle = rs.getBoolean("clean_title");
                 int licenseNum = rs.getInt("license_num");
-                double price = rs.getInt("price");
+                double price = rs.getDouble("price");
                 Car dbCar = new Car(vin, dbMake, dbModel, modelYear, color, cleanTitle, licenseNum, price);
                 carList.add(dbCar);
             }
@@ -99,7 +99,7 @@ public class CarDAO {
                 String carColor = rs.getString("color");
                 boolean cleanTitle = rs.getBoolean("clean_title");
                 int licenseNum = rs.getInt("license_num");
-                double price = rs.getInt("price");
+                double price = rs.getDouble("price");
                 Car dbCar = new Car(vin, dbMake, dbModel, modelYear, carColor, cleanTitle, licenseNum, price);
                 carList.add(dbCar);
             }
@@ -123,7 +123,7 @@ public class CarDAO {
                 String color = rs.getString("color");
                 boolean cleanTitle = rs.getBoolean("clean_title");
                 int licenseNum = rs.getInt("license_num");
-                double price = rs.getInt("price");
+                double price = rs.getDouble("price");
                 Car dbCar = new Car(vin, dbMake, dbModel, modelYear, color, cleanTitle, licenseNum, price);
                 carList.add(dbCar);
             }
@@ -173,7 +173,7 @@ public class CarDAO {
                 String color = rs.getString("color");
                 boolean cleanTitle = rs.getBoolean("clean_title");
                 int licenseNum = rs.getInt("license_num");
-                double price = rs.getInt("price");
+                double price = rs.getDouble("price");
                 Car dbCar = new Car(vin, dbMake, dbModel, modelYear, color, cleanTitle, licenseNum, price);
                 carList.add(dbCar);
             }
